Refactor : 로그인시 profile img정보 response에 추가
