+++ src/main/java/reduck/reduck/domain/auth/controller/AuthController.java
@@ -6,8 +6,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import reduck.reduck.domain.auth.dto.AccessTokenDto;
 import reduck.reduck.domain.auth.service.AuthService;
-import reduck.reduck.domain.user.dto.SignInDto;
-import reduck.reduck.domain.user.dto.SignInResponseDto;
+import reduck.reduck.domain.auth.dto.SignInDto;
+import reduck.reduck.domain.auth.dto.SignInResponseDto;
 
 import javax.servlet.http.HttpServletRequest;
