+++ src/main/java/reduck/reduck/domain/auth/dto/SignInDto.java
@@ -1,4 +1,4 @@
-package reduck.reduck.domain.user.dto;
+package reduck.reduck.domain.auth.dto;
 
 import lombok.Getter;
 import lombok.Setter;
