+++ src/main/java/reduck/reduck/domain/auth/dto/SignInResponseDto.java
@@ -1,14 +1,14 @@
-package reduck.reduck.domain.user.dto;
+package reduck.reduck.domain.auth.dto;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.*;
 import reduck.reduck.domain.user.entity.Authority;
+import reduck.reduck.domain.user.entity.UserProfileImg;
+
 import java.util.ArrayList;
 import java.util.List;
 
 @Getter
+@Setter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
@@ -22,8 +22,10 @@ public class SignInResponseDto {
     private String email;
 
     private List<Authority> roles = new ArrayList<>();
+    private UserProfileImg userProfileImg;
 
     private String accessToken;
     private String refreshToken;
+
 
 }
